requirements.txt
flask
flask-cors
Pillow
torch
timm
torchvision
flask-limiter
python-dotenv
python-multipart
